3_anime_fan/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.5
django-mongodb-engine @ git+https://github.com/django-nonrel/mongodb-engine@1314786975d3d17fb1a2e870dfba004158a5e431
djangorestframework==3.13.1
djangotoolbox @ git+https://github.com/django-nonrel/djangotoolbox@7e2c47e2ab5ee98d0c0f47c5ccce08e75feda48a
djongo==1.3.6
ffmpeg-python==0.2.0
future==0.18.2
kombu==5.2.4
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pymongo==3.12.3
python-ffmpeg-video-streaming==0.1.15
pytz==2022.1
six==1.16.0
sqlparse==0.2.4
vine==5.0.0
wcwidth==0.2.5
